+++ src/store/index.js
@@ -10,7 +10,7 @@ import {account, setAuthInHeader} from '../api'
 //     deleteQuestion, deleteAddress
 // } from '../api/index'
 import {
-    fetchAccountCarts, deleteCartItem, addOrderItems,
+    addOrderItems,
     addCupon, fetchCuponsByAccount,
     addAdress, fetchAdressByAccount, deleteAddress
 } from '../api/index'
@@ -18,6 +18,7 @@ import products from "./modules/products";
 import reviews from "./modules/reviews";
 import questions from "./modules/questions";
 import orders from "./modules/orders";
+import carts from "./modules/carts";
 
 Vue.use(Vuex);
 
@@ -26,15 +27,13 @@ const store = new Vuex.Store({
         reviews,
         products,
         questions,
-        orders
+        orders,
+        carts
     },
     state: {
         account: {},
         accessToken: null,
-        carts: {},
-        // orderInfo: {},
-        // payInfo: {},
-        // orderLog: {},
+        // carts: {},
         myAddresses: {},
         myCupons: {}
     },
@@ -51,21 +50,12 @@ const store = new Vuex.Store({
             delete localStorage.accessToken;
             setAuthInHeader(null);
         },
-        SET_CARTS(state, carts) {
-            state.carts = carts;
-        },
-        // SET_ORDER_INFO(state, orderInfo) {
-        //     state.orderInfo = orderInfo;
-        // },
-        // SET_PAY_INFO(state, payInfo) {
-        //     state.payInfo = payInfo;
+        // SET_CARTS(state, carts) {
+        //     state.carts = carts;
         // },
         SET_CUPONS(state, cupons) {
             state.orderInfo.accountCupons = cupons;
         },
-        // SET_ORDER_LOG(state, orderLog) {
-        //     state.orderLog = orderLog;
-        // },
         SET_MY_ADDRESSES(state, myAddresses) {
             state.myAddresses = myAddresses;
         },
@@ -79,28 +69,19 @@ const store = new Vuex.Store({
             commit('LOGIN', response.data);
             return response;
         },
-        // async FETCH_ORDERLOG(_, payload) {
-        //     const response = await fetchOrderLog(payload);
+        // async FETCH_ACCOUNT_CARTS({commit}) {
+        //     const response = await fetchAccountCarts();
+        //     commit('SET_CARTS', response.data);
         //     return response;
         // },
-        async FETCH_ACCOUNT_CARTS({commit}) {
-            const response = await fetchAccountCarts();
-            commit('SET_CARTS', response.data);
-            return response;
-        },
-        async DELETE_CART_ITEM(_, payload) {
-            const response = await deleteCartItem(payload.cartId);
-            return response;
-        },
+        // async DELETE_CART_ITEM(_, payload) {
+        //     const response = await deleteCartItem(payload.cartId);
+        //     return response;
+        // },
         async ADD_ORDER_ITEMS(_, payload) {
             const response = await addOrderItems(payload);
             return response;
         },
-        // async FETCH_ORDER_INFO({commit}, payload) {
-        //     const response = await fetchOrderInfo(payload);
-        //     commit('SET_ORDER_INFO', response.data);
-        //     return response;
-        // },
         async ADD_CUPON(_, payload) {
             const response = await addCupon(payload);
             return response;
@@ -117,20 +98,6 @@ const store = new Vuex.Store({
             const response = await fetchAdressByAccount(payload);
             return response;
         },
-        // async ADD_PAY_ORDER(_, payload) {
-        //     const response = await addPayOrder(payload);
-        //     return response;
-        // },
-        // async FETCH_MY_ORDER_LOG({commit}, payload) {
-        //     const response = await fetchMyOrderLog(payload);
-        //     commit('SET_ORDER_LOG', response.data);
-        //     return response;
-        // },
-        // async CANCEL_ORDER(_, payload) {
-        //     const response = await cancelOrder(payload);
-        //     return response;
-        // },
-
         async DELETE_ADDRESS(_, payload) {
             const response = await deleteAddress(payload);
             return response;
